pyproject.toml
[tool.black]
line-length = 72
py37 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
| dist
)/
'''

[tool.poetry]
name = "media_scraper"
version = "0.1.0"
description = ""
authors = ["Andrés Sandoval <andres.sandoval@alumnos.usm.cl>"]

[tool.poetry.dependencies]
python = "^3.6"
structlog = "^20.1.0"
python-dotenv = "^0.14.0"
lxml = "^4.5.1"
requests = "^2.24.0"
requests_html = "^0.10.0"
facebook-scraper = "^0.2.3"
elastinga = {path = "elastinga-0.1.0-py3-none-any.whl"}

[tool.poetry.dev-dependencies]
pytest = "^4.6"
pylint = "^2.5.3"
ipython = "^7.16.1"
black = {version = "^19.10b0", allow-prereleases = true}
colorama = "^0.4.3"
autopep8 = "^1.5.3"
docformatter = "^1.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
